feat: move helper assignment logic to a separate module for better organization and maintainability
